Moved INPC model perf tests to sub folder
