+++ Model/Tests/RxCookbook/INPC/PropertyExtensions.cs
@@ -1,0 +1,33 @@
+using System;
+using System.Linq.Expressions;
+using System.Reflection;
+
+namespace RxCookbook.INPC
+{
+    public static class PropertyExtensions
+    {
+        /// <summary>
+        /// Gets property information for the specified <paramref name="property"/> expression.
+        /// </summary>
+        /// <typeparam name="TSource">Type of the parameter in the <paramref name="property"/> expression.</typeparam>
+        /// <typeparam name="TValue">Type of the property's value.</typeparam>
+        /// <param name="property">The expression from which to retrieve the property information.</param>
+        /// <returns>Property information for the specified expression.</returns>
+        /// <exception cref="ArgumentException">The expression is not understood.</exception>
+        public static PropertyInfo GetPropertyInfo<TSource, TValue>(this Expression<Func<TSource, TValue>> property)
+        {
+            if (property == null)
+                throw new ArgumentNullException("property");
+
+            var body = property.Body as MemberExpression;
+            if (body == null)
+                throw new ArgumentException("Expression is not a property", "property");
+
+            var propertyInfo = body.Member as PropertyInfo;
+            if (propertyInfo == null)
+                throw new ArgumentException("Expression is not a property", "property");
+
+            return propertyInfo;
+        }
+    }
+}
